+++ test_project1/src/main/java/com/guodong/core/service/EmployeeService.java
@@ -10,7 +10,7 @@ public interface EmployeeService {
 
 	void updateEmployee(Employee employee);
 
-	void deleteEmployeeWithId(Integer[] ids);
+	void deleteEmployeeWithId(Long[] ids);
 
 	Integer getAllEmployee();
 }
